+++ src/Input.cpp
@@ -14,7 +14,7 @@ OutputInfo Input::initialize(std::vector<double*>& parameterPointers,
 {
   // Bias component will not change after initialization
   if(bias)
-    y(J) = double(1.0);
+    y(J) = 1.0;
   OutputInfo info;
   info.bias = bias;
   info.dimensions.push_back(dim1);
@@ -38,7 +38,7 @@ void Input::forwardPropagate(Eigen::VectorXd* x, Eigen::VectorXd*& y, bool dropo
     RandomNumberGenerator rng;
     for(int j = 0; j < J; j++)
       if(rng.generate<double>(0.0, 1.0) < dropoutProbability)
-        this->y(j) = (double) 0;
+        this->y(j) = 0.0;
   }
   else if(dropoutProbability > 0.0)
   {
